cdk/requirements.txt
aws-cdk-lib==2.99.1
constructs==10.3.0
boto3>=1.26.0